Added CORS support for localhost:3000
So that data can be requested from the dev front-end, CORS support for the origin "http://localhost:3000" has been added. It's not ideal, as it means anybody hosting on that origin can use our API, but it's not a big deal since they could do that via our site anyway.